trying to come up with the best abstraction for the display

+++ Emulator.Tests/Fakes/FakeDisplay.cs
@@ -10,7 +10,7 @@ internal sealed class FakeDisplay : IDisplay
     public void Clear() => ClearCount++;
     public void Update() => UpdateCount++;
     
-    public int BlitRow(byte x, byte y, byte row)
+    public byte BlitRow(byte x, byte y, byte row)
     {
         throw new NotImplementedException();
     }
